refactor: move chip-display into backend subcommand, remove separate chip CLI
`uniqc backend chip-display platform/chip_name` now shows per-qubit chip
characterization data (T1/T2, gate fidelity, readout fidelity, connectivity).
The underlying data layer (chip_info.py / chip_cache.py / chip_service.py)
remains unchanged.
