+++ uniqc/cli/backend.py
@@ -32,7 +32,7 @@ app = typer.Typer(
 
 # Detect whether a subcommand was given by checking if the first
 # positional arg in sys.argv matches a known subcommand name.
-_SUB_COMMANDS = frozenset({"list", "update", "show"})
+_SUB_COMMANDS = frozenset({"list", "update", "show", "chip-display"})
 
 
 def _subcommand_given() -> bool:
@@ -412,6 +412,153 @@ def _print_backend_detail(b: BackendInfo) -> None:
             else:
                 display = str(value)
             console.print(f"  [cyan]{key}:[/cyan] {display}")
+
+
+@app.command("chip-display")
+def chip_display(
+    identifier: str = typer.Argument(
+        ...,
+        help="Backend identifier in the form 'platform/chip_name' (e.g. originq/wuyuan:d5, ibm/sherbrooke, quafu/ScQ-P18)",
+    ),
+    update: bool = typer.Option(
+        False,
+        "--update",
+        "-u",
+        help="Force-refresh chip data from the cloud before displaying",
+    ),
+    ai_hints: bool = AI_HINTS_OPTION,
+):
+    """Show per-qubit chip characterization data for a backend.
+
+    Displays detailed calibration data including: T1/T2 coherence times,
+    single-qubit gate fidelity, readout fidelity (R0/R1/avg), and
+    two-qubit gate fidelity per connected pair.
+
+    The chip data is cached locally after first fetch. Use --update to
+    force-refresh from the cloud.
+
+    Workflow:
+      - Pick a backend from: uniqc backend list
+      - Force refresh if calibration has been updated: uniqc backend chip-display originq/wuyuan:d5 --update
+      - Use this data for qubit selection: see the analyzer module.
+    """
+    if ai_hints or os.environ.get("UNIQC_AI_HINTS"):
+        print_ai_hints("backend-chip-display")
+
+    parts = identifier.split("/", 1)
+    if len(parts) != 2:
+        print_error("Identifier must be in the form 'platform/chip_name', e.g. originq/wuyuan:d5")
+        raise typer.Exit(1) from None
+
+    platform_str, chip_name = parts
+    try:
+        platform = Platform(platform_str.strip().lower())
+    except ValueError:
+        print_error(f"Unknown platform '{platform_str}'. Valid: originq, quafu, ibm")
+        raise typer.Exit(1) from None
+
+    from uniqc.chip_service import fetch_chip_characterization
+
+    chip = fetch_chip_characterization(chip_name.strip(), platform, force_refresh=update)
+
+    if chip is None:
+        print_error(
+            f"Chip '{chip_name}' not found on platform '{platform_str}' "
+            "or the platform is unavailable (check credentials with: uniqc config validate)"
+        )
+        raise typer.Exit(1) from None
+
+    _print_chip_detail(chip)
+
+
+def _fmt(val: float | None, pattern: str) -> str:
+    """Format a float or return '-' if None."""
+    if val is None:
+        return "-"
+    return f"{val:{pattern}}"
+
+
+def _print_chip_detail(chip) -> None:
+    """Print chip characterization details using Rich formatting."""
+    from rich.panel import Panel
+    from rich.rule import Rule
+    from rich.table import Table
+
+    console.print(Rule(f"[bold cyan]Chip: {chip.full_id}[/bold cyan]"))
+
+    # Overview panel
+    overview = [
+        f"[bold]Platform:[/bold]  {chip.platform.value}",
+        f"[bold]Chip:[/bold]      {chip.chip_name}",
+        f"[bold]Qubits:[/bold]    {len(chip.available_qubits)} available / {len(chip.connectivity)} pairs",
+        f"[bold]Calibrated:[/bold] {chip.calibrated_at or 'N/A'}",
+    ]
+    console.print(Panel("\n".join(overview), title="Overview", box=rich.box.ROUNDED))
+
+    # Global info
+    gi = chip.global_info
+    if gi.single_qubit_gates or gi.two_qubit_gates or gi.single_qubit_gate_time:
+        gi_lines: list[str] = []
+        if gi.single_qubit_gates:
+            gi_lines.append(f"[bold]1Q Gates:[/bold]     {', '.join(gi.single_qubit_gates)}")
+        if gi.two_qubit_gates:
+            gi_lines.append(f"[bold]2Q Gates:[/bold]     {', '.join(gi.two_qubit_gates)}")
+        if gi.single_qubit_gate_time:
+            gi_lines.append(f"[bold]1Q Gate Time:[/bold] {gi.single_qubit_gate_time} ns")
+        if gi.two_qubit_gate_time:
+            gi_lines.append(f"[bold]2Q Gate Time:[/bold] {gi.two_qubit_gate_time} ns")
+        console.print(Panel("\n".join(gi_lines), title="Global Info", box=rich.box.ROUNDED))
+
+    # Per-qubit table
+    if chip.single_qubit_data:
+        q_table = Table(
+            title="Per-Qubit Data",
+            box=rich.box.ROUNDED,
+            show_header=True,
+            header_style="bold cyan",
+        )
+        q_table.add_column("ID", justify="right", width=5)
+        q_table.add_column("T1 (μs)", justify="right", width=9)
+        q_table.add_column("T2 (μs)", justify="right", width=9)
+        q_table.add_column("1Q Fid.", justify="right", width=9)
+        q_table.add_column("R0", justify="right", width=7)
+        q_table.add_column("R1", justify="right", width=7)
+        q_table.add_column("Avg R", justify="right", width=9)
+
+        for sq in sorted(chip.single_qubit_data, key=lambda x: x.qubit_id):
+            q_table.add_row(
+                str(sq.qubit_id),
+                _fmt(sq.t1, ".2f"),
+                _fmt(sq.t2, ".2f"),
+                _fmt(sq.single_gate_fidelity, ".4f"),
+                _fmt(sq.readout_fidelity_0, ".4f"),
+                _fmt(sq.readout_fidelity_1, ".4f"),
+                _fmt(sq.avg_readout_fidelity, ".4f"),
+            )
+        console.print(q_table)
+
+    # Per-pair table
+    if chip.two_qubit_data:
+        p_table = Table(
+            title="Per-Pair 2Q Gate Data",
+            box=rich.box.ROUNDED,
+            show_header=True,
+            header_style="bold cyan",
+        )
+        p_table.add_column("Qubit U", justify="right", width=8)
+        p_table.add_column("Qubit V", justify="right", width=8)
+        p_table.add_column("Gate", width=8)
+        p_table.add_column("Fidelity", justify="right", width=10)
+
+        for tp in sorted(chip.two_qubit_data, key=lambda x: (x.qubit_u, x.qubit_v)):
+            for gate in tp.gates:
+                p_table.add_row(
+                    str(tp.qubit_u),
+                    str(tp.qubit_v),
+                    gate.gate,
+                    _fmt(gate.fidelity, ".4f"),
+                )
+        console.print(p_table)
 
 
 def _format_age(seconds: float) -> str:
